Изображения: модель и миграции (полиморфные отношения)

+++ database/migrations/2025_01_17_200227_create_notifications_table.php
@@ -4,12 +4,12 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
-    public function up(): void
+    public function up()
+    : void
     {
         Schema::create('notifications', function (Blueprint $table) {
             $table->uuid('id')->primary();
@@ -24,8 +24,11 @@ return new class extends Migration
     /**
      * Reverse the migrations.
      */
-    public function down(): void
+    public function down()
+    : void
     {
-        Schema::dropIfExists('notifications');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('notifications');
+        }
     }
 };
